+++ nupic/research/frameworks/vernon/__init__.py
@@ -20,5 +20,3 @@
 
 from .common_experiments import *
 from .handlers import *
-from .run_experiment.run import run, terminate_processes
-from .run_experiment.run_with_raytune import run as run_with_raytune
