+++ nupic/research/frameworks/vernon/run_with_raytune.py
@@ -29,12 +29,14 @@ from ray.tune import Trainable, tune
 
 from nupic.research.frameworks.vernon.experiment_utils import get_free_port
 from nupic.research.frameworks.vernon.search import TrialsCollection
+from nupic.research.frameworks.vernon.trainables import (
+    SigOptSupervisedTrainable,
+    SupervisedTrainable,
+)
 from nupic.research.support.ray_utils import (
     get_last_checkpoint,
     register_torch_serializers,
 )
-
-from .trainables import SigOptSupervisedTrainable, SupervisedTrainable
 
 os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE"
