+++ projects/gsc/run.py
@@ -21,8 +21,8 @@ import argparse
 import copy
 
 from experiments import CONFIGS
-from nupic.research.frameworks import vernon
 from nupic.research.frameworks.vernon.parser_utils import DEFAULT_PARSERS, process_args
+from nupic.research.frameworks.vernon.run_with_raytune import run
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
@@ -48,4 +48,4 @@ if __name__ == "__main__":
     if config is None:
         pass
     else:
-        vernon.run_with_raytune(config)
+        run(config)
